*Issue #, if available:*
Perf tests are all on small payloads, and we can't commit large payloads. And besides, creating test data can be laborious, especially with different content types, dimensionality, and payload sizes.

*Description of changes:*

Adds a script that takes in content type, input shape, and payload size and prints to stdout the output. Adds some additional perftests on 10MB payloads with 10000 dimensions on the half_plus_three model.

Results below include the changes in `https://github.com/aws/sagemaker-tensorflow-serving-container/pull/28`. They are for CSV, JSON, and JSONLines, respectively:

```
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:        nginx/1.14.2
Server Hostname:        localhost
Server Port:            8080

Document Path:          /invocations
Document Length:        26251078 bytes

Concurrency Level:      1
Time taken for tests:   23.874 seconds
Complete requests:      10
Failed requests:        0
Keep-Alive requests:    10
Total transferred:      262512380 bytes
Total body sent:        105001680
HTML transferred:       262510780 bytes
Requests per second:    0.42 [#/sec] (mean)
Time per request:       2387.411 [ms] (mean)
Time per request:       2387.411 [ms] (mean, across all concurrent requests)
Transfer rate:          10737.98 [Kbytes/sec] received
                        4295.06 kb/s sent
                        15033.04 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       0
Processing:  2056 2387 452.7   2077    3163
Waiting:     1503 1657 212.8   1576    2088
Total:       2056 2387 452.7   2077    3163

Percentage of the requests served within a certain time (ms)
  50%   2077
  66%   2569
  75%   2632
  80%   3115
  90%   3163
  95%   3163
  98%   3163
  99%   3163
 100%   3163 (longest request)
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:        nginx/1.14.2
Server Hostname:        localhost
Server Port:            8080

Document Path:          /invocations
Document Length:        17500728 bytes

Concurrency Level:      1
Time taken for tests:   13.967 seconds
Complete requests:      10
Failed requests:        0
Keep-Alive requests:    10
Total transferred:      175008880 bytes
Total body sent:        105008770
HTML transferred:       175007280 bytes
Requests per second:    0.72 [#/sec] (mean)
Time per request:       1396.717 [ms] (mean)
Time per request:       1396.717 [ms] (mean, across all concurrent requests)
Transfer rate:          12236.34 [Kbytes/sec] received
                        7342.05 kb/s sent
                        19578.39 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:  1310 1397  75.2   1398    1567
Waiting:     1006 1044  35.2   1044    1116
Total:       1310 1397  75.2   1398    1567

Percentage of the requests served within a certain time (ms)
  50%   1398
  66%   1416
  75%   1427
  80%   1444
  90%   1567
  95%   1567
  98%   1567
  99%   1567
 100%   1567 (longest request)
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:        nginx/1.14.2
Server Hostname:        localhost
Server Port:            8080

Document Path:          /invocations
Document Length:        17500728 bytes

Concurrency Level:      1
Time taken for tests:   15.170 seconds
Complete requests:      10
Failed requests:        0
Keep-Alive requests:    10
Total transferred:      175008880 bytes
Total body sent:        105005310
HTML transferred:       175007280 bytes
Requests per second:    0.66 [#/sec] (mean)
Time per request:       1516.995 [ms] (mean)
Time per request:       1516.995 [ms] (mean, across all concurrent requests)
Transfer rate:          11266.16 [Kbytes/sec] received
                        6759.69 kb/s sent
                        18025.85 kb/s total

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:  1382 1517 139.9   1489    1838
Waiting:     1008 1073  74.4   1080    1227
Total:       1382 1517 139.8   1489    1838

Percentage of the requests served within a certain time (ms)
  50%   1489
  66%   1513
  75%   1532
  80%   1666
  90%   1838
  95%   1838
  98%   1838
  99%   1838
 100%   1838 (longest request)
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.